// packages/hashpin-contracts/typechain-types/Hashpin.ts
export interface HashpinInterface extends Interface {
  getFunction(
    nameOrSignature:
      | "INITIAL_DIFFICULTY"
      | "currentDifficulty"
      | "getDifficulty"
      | "getHashDetails"
      | "meetsDifficulty"
      | "pinHash"
      | "pinnedHashes"
      | "setDifficulty"
      | "verifyHash"
  ): FunctionFragment;

  getEvent(
    nameOrSignatureOrTopic: "DifficultyUpdated" | "HashPinned"
  ): EventFragment;

  encodeFunctionData(
    functionFragment: "INITIAL_DIFFICULTY",
    values?: undefined
  ): string;
  encodeFunctionData(
    functionFragment: "currentDifficulty",
    values?: undefined
  ): string;
  encodeFunctionData(
    functionFragment: "getDifficulty",
    values?: undefined
  ): string;
  encodeFunctionData(
    functionFragment: "getHashDetails",
    values: [BytesLike]
  ): string;
  encodeFunctionData(
    functionFragment: "meetsDifficulty",
    values: [BytesLike]
  ): string;
  encodeFunctionData(
    functionFragment: "pinHash",
    values: [BytesLike, BigNumberish, string]
  ): string;
  encodeFunctionData(
    functionFragment: "pinnedHashes",
    values: [BytesLike]
  ): string;
  encodeFunctionData(
    functionFragment: "setDifficulty",
    values: [BigNumberish]
  ): string;
  encodeFunctionData(
    functionFragment: "verifyHash",
    values: [BytesLike]
  ): string;

  decodeFunctionResult(
    functionFragment: "INITIAL_DIFFICULTY",
    data: BytesLike
  ): Result;
  decodeFunctionResult(
    functionFragment: "currentDifficulty",
    data: BytesLike
  ): Result;
  decodeFunctionResult(
    functionFragment: "getDifficulty",
    data: BytesLike
  ): Result;
  decodeFunctionResult(
    functionFragment: "getHashDetails",
    data: BytesLike
  ): Result;
  decodeFunctionResult(
    functionFragment: "meetsDifficulty",
    data: BytesLike
  ): Result;
  decodeFunctionResult(functionFragment: "pinHash", data: BytesLike): Result;
  decodeFunctionResult(
    functionFragment: "pinnedHashes",
    data: BytesLike
  ): Result;
  decodeFunctionResult(
    functionFragment: "setDifficulty",
    data: BytesLike
  ): Result;
  decodeFunctionResult(functionFragment: "verifyHash", data: BytesLike): Result;
}

export namespace DifficultyUpdatedEvent {
  export type InputTuple = [newDifficulty: BigNumberish];
  export type OutputTuple = [newDifficulty: bigint];
  export interface OutputObject {
    newDifficulty: bigint;
  }
  export type Event = TypedContractEvent<InputTuple, OutputTuple, OutputObject>;
  export type Filter = TypedDeferredTopicFilter<Event>;
  export type Log = TypedEventLog<Event>;
  export type LogDescription = TypedLogDescription<Event>;
}

export namespace HashPinnedEvent {
  export type InputTuple = [
    pinner: AddressLike,
    hash: BytesLike,
    metadata: string,
    timestamp: BigNumberish
  ];
  export type OutputTuple = [
    pinner: string,
    hash: string,
    metadata: string,
    timestamp: bigint
  ];
  export interface OutputObject {
    pinner: string;
    hash: string;
    metadata: string;
    timestamp: bigint;
  }
  export type Event = TypedContractEvent<InputTuple, OutputTuple, OutputObject>;
  export type Filter = TypedDeferredTopicFilter<Event>;
  export type Log = TypedEventLog<Event>;
  export type LogDescription = TypedLogDescription<Event>;
}

export interface Hashpin extends BaseContract {
  connect(runner?: ContractRunner | null): Hashpin;
  waitForDeployment(): Promise<this>;

  interface: HashpinInterface;

  queryFilter<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    fromBlockOrBlockhash?: string | number | undefined,
    toBlock?: string | number | undefined
  ): Promise<Array<TypedEventLog<TCEvent>>>;
  queryFilter<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    fromBlockOrBlockhash?: string | number | undefined,
    toBlock?: string | number | undefined
  ): Promise<Array<TypedEventLog<TCEvent>>>;

  on<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    listener: TypedListener<TCEvent>
  ): Promise<this>;
  on<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    listener: TypedListener<TCEvent>
  ): Promise<this>;

  once<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    listener: TypedListener<TCEvent>
  ): Promise<this>;
  once<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    listener: TypedListener<TCEvent>
  ): Promise<this>;

  listeners<TCEvent extends TypedContractEvent>(
    event: TCEvent
  ): Promise<Array<TypedListener<TCEvent>>>;
  listeners(eventName?: string): Promise<Array<Listener>>;
  removeAllListeners<TCEvent extends TypedContractEvent>(
    event?: TCEvent
  ): Promise<this>;

  INITIAL_DIFFICULTY: TypedContractMethod<[], [bigint], "view">;

  currentDifficulty: TypedContractMethod<[], [bigint], "view">;

  getDifficulty: TypedContractMethod<[], [bigint], "view">;

  getHashDetails: TypedContractMethod<
    [hash: BytesLike],
    [
      [string, string, bigint] & {
        pinner: string;
        metadata: string;
        timestamp: bigint;
      }
    ],
    "view"
  >;

  meetsDifficulty: TypedContractMethod<[hash: BytesLike], [boolean], "view">;

  pinHash: TypedContractMethod<
    [hash: BytesLike, nonce: BigNumberish, metadata: string],
    [void],
    "nonpayable"
  >;

  pinnedHashes: TypedContractMethod<
    [arg0: BytesLike],
    [
      [string, string, string, bigint] & {
        pinner: string;
        hash: string;
        metadata: string;
        timestamp: bigint;
      }
    ],
    "view"
  >;

  setDifficulty: TypedContractMethod<
    [newDifficulty: BigNumberish],
    [void],
    "nonpayable"
  >;

  verifyHash: TypedContractMethod<[hash: BytesLike], [string], "view">;

  getFunction<T extends ContractMethod = ContractMethod>(
    key: string | FunctionFragment
  ): T;

  getFunction(
    nameOrSignature: "INITIAL_DIFFICULTY"
  ): TypedContractMethod<[], [bigint], "view">;
  getFunction(
    nameOrSignature: "currentDifficulty"
  ): TypedContractMethod<[], [bigint], "view">;
  getFunction(
    nameOrSignature: "getDifficulty"
  ): TypedContractMethod<[], [bigint], "view">;
  getFunction(
    nameOrSignature: "getHashDetails"
  ): TypedContractMethod<
    [hash: BytesLike],
    [
      [string, string, bigint] & {
        pinner: string;
        metadata: string;
        timestamp: bigint;
      }
    ],
    "view"
  >;
  getFunction(
    nameOrSignature: "meetsDifficulty"
  ): TypedContractMethod<[hash: BytesLike], [boolean], "view">;
  getFunction(
    nameOrSignature: "pinHash"
  ): TypedContractMethod<
    [hash: BytesLike, nonce: BigNumberish, metadata: string],
    [void],
    "nonpayable"
  >;
  getFunction(
    nameOrSignature: "pinnedHashes"
  ): TypedContractMethod<
    [arg0: BytesLike],
    [
      [string, string, string, bigint] & {
        pinner: string;
        hash: string;
        metadata: string;
        timestamp: bigint;
      }
    ],
    "view"
  >;
  getFunction(
    nameOrSignature: "setDifficulty"
  ): TypedContractMethod<[newDifficulty: BigNumberish], [void], "nonpayable">;
  getFunction(
    nameOrSignature: "verifyHash"
  ): TypedContractMethod<[hash: BytesLike], [string], "view">;

  getEvent(
    key: "DifficultyUpdated"
  ): TypedContractEvent<
    DifficultyUpdatedEvent.InputTuple,
    DifficultyUpdatedEvent.OutputTuple,
    DifficultyUpdatedEvent.OutputObject
  >;
  getEvent(
    key: "HashPinned"
  ): TypedContractEvent<
    HashPinnedEvent.InputTuple,
    HashPinnedEvent.OutputTuple,
    HashPinnedEvent.OutputObject
  >;

  filters: {
    "DifficultyUpdated(uint256)": TypedContractEvent<
      DifficultyUpdatedEvent.InputTuple,
      DifficultyUpdatedEvent.OutputTuple,
      DifficultyUpdatedEvent.OutputObject
    >;
    DifficultyUpdated: TypedContractEvent<
      DifficultyUpdatedEvent.InputTuple,
      DifficultyUpdatedEvent.OutputTuple,
      DifficultyUpdatedEvent.OutputObject
    >;

    "HashPinned(address,bytes32,string,uint256)": TypedContractEvent<
      HashPinnedEvent.InputTuple,
      HashPinnedEvent.OutputTuple,
      HashPinnedEvent.OutputObject
    >;
    HashPinned: TypedContractEvent<
      HashPinnedEvent.InputTuple,
      HashPinnedEvent.OutputTuple,
      HashPinnedEvent.OutputObject
    >;
  };
}